consolidate metadata and db to preprocess
